+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./brand";
+export * from "./format";
 export * from "./types";
 export * from "./auth";
 export {